+++ crates/gpui/src/platform.rs
@@ -1681,11 +1681,3 @@ impl From<String> for ClipboardString {
         }
     }
 }
-
-/// A trait for platform-specific keyboard layouts
-pub trait PlatformKeyboardLayout {
-    /// Get the keyboard layout ID, which should be unique to the layout
-    fn id(&self) -> &str;
-    /// Get the keyboard layout display name
-    fn name(&self) -> &str;
-}
