+++ crates/gpui/src/platform/windows.rs
@@ -4,6 +4,7 @@ mod direct_write;
 mod dispatcher;
 mod display;
 mod events;
+mod keyboard;
 mod platform;
 mod system_settings;
 mod util;
@@ -16,6 +17,7 @@ pub(crate) use direct_write::*;
 pub(crate) use dispatcher::*;
 pub(crate) use display::*;
 pub(crate) use events::*;
+pub(crate) use keyboard::*;
 pub(crate) use platform::*;
 pub(crate) use system_settings::*;
 pub(crate) use util::*;
